[Refactor]
Extract picker extension
