+++ Sources/Rings/ArchimedeanSpiralText.swift
@@ -151,41 +151,6 @@ extension ArchimedeanSpiralText: Adjustable {
     }
 }
 
-struct SegmentedPicker: ViewModifier {
-    func body(content: Content) -> some View {
-        #if os(watchOS)
-        content
-        #else
-        content.pickerStyle(SegmentedPickerStyle()).padding(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 7))
-        #endif
-    }
-}
-
-struct ColoredPicker: ViewModifier {
-    @Binding var selection: Color
-    func body(content: Content) -> some View {
-        #if os(macOS) || os(iOS)
-        if #available(macOS 11.0, iOS 14.0, macCatalyst 14.0, *) {
-            ColorPicker("", selection: _selection)
-        } else {
-            content.modifier(SegmentedPicker())
-        }
-        #else
-        content.modifier(SegmentedPicker())
-        #endif
-    }
-}
-
-extension Picker {
-    func segmented() -> some View {
-        modifier(SegmentedPicker())
-    }
-    
-    func colorPicker(_ selection: Binding<Color>) -> some View {
-        modifier(ColoredPicker(selection: selection))
-    }
-}
-
 #if !os(tvOS)
 struct ArchimedeanSpiralText_Preview_macOS : View {
     private let demoText = "1234567890abcdefgABCDEFG♩♪♫♬"
